[4/N] Further refactor FR script to make it more modulized
[ghstack-poisoned]
