+++ test/distributed/flight_recorder/test_fr_analysis.py
@@ -7,7 +7,8 @@ import sys
 REPO_ROOT = pathlib.Path(__file__).resolve().parent.parent.parent.parent
 
 sys.path.insert(0, str(REPO_ROOT))
-from tools.flight_recorder.components.processor import match_one_event, MatchState
+from tools.flight_recorder.components.types import MatchState
+from tools.flight_recorder.components.utils import match_one_event
 
 
 # Make sure to remove REPO_ROOT after import is done
