+++ tools/flight_recorder/components/builder.py
@@ -22,6 +22,10 @@ from typing import (  # type: ignore[attr-defined]
     TypeVar,
 )
 
+from tools.flight_recorder.components.utils import match_one_event, match_coalesced_groups, find_coalesced_group, just_print_entries, check_no_missing_dump_files, check_version, check_trace_from_beginning
+from tools.flight_recorder.components.types import Op, Group, Membership, NCCLCall, Traceback, Collective, MatchState, Database
+
+
 
 try:
     from tabulate import tabulate
@@ -33,433 +37,9 @@ except ModuleNotFoundError:
         return data
 
 
-T = TypeVar("T", bound=NamedTuple)
-
-
-class Ref(Generic[T]):
-    pass
-
-
-class TypeInfo(NamedTuple):
-    name: str
-    fields: List[Tuple[str, Type]]  # type: ignore[type-arg]
-
-    @classmethod
-    def from_type(cls, c: T) -> "TypeInfo":
-        if hasattr(c, "__name__"):
-            name = c.__name__
-        else:
-            name = str(c)
-        return cls(
-            name,
-            [(f, _eval_type(c.__annotations__[f], globals(), {})) for f in c._fields],
-        )
-
-
-"""
-Schema for flat DB
-
-TODO schemas not yet implemented
-# threads as recorded at termination of process
-Threads
-    id: int
-    traceback_id: int
-    process_id: int
-
-Process:
-    id: int # Same as world groups RANK
-    pid: int
-    hostname: str
-
-NCCLOp:
-    # nccl op implementation details (sends/recv)
-    id: int
-    nccl_call_id: int
-
-"""
-
-
-class Group(NamedTuple):
-    id: int
-    desc: str
-    size: int
-
-
-class Membership(NamedTuple):
-    group_id: Ref[Group]
-    global_rank: int
-
-
-class Traceback(NamedTuple):
-    id: int
-    frames: str
-
-
-class Collective(NamedTuple):
-    id: int
-    group_id: Ref[Group]
-
-
-class NCCLCall(NamedTuple):
-    id: int
-    collective_id: Ref[Collective]
-    group_id: Ref[Group]
-    global_rank: int  # technically Ref[Process] once we have it
-    traceback_id: Ref[Traceback]
-    collective_type: str
-    sizes: List[List[int]]
-
-
-class Database(NamedTuple):
-    groups: List[Group]
-    memberships: List[Membership]
-    tracebacks: List[Traceback]
-    collectives: List[Collective]
-    ncclcalls: List[NCCLCall]
-
-
-# TODO: We need to add a schema for the following
-types = [
-    TypeInfo.from_type(t)  # type: ignore[type-var]
-    for t in globals().values()
-    if (
-        isinstance(t, type)
-        and issubclass(t, tuple)
-        and hasattr(t, "_fields")
-        and t is not TypeInfo
-    )
-]
-
-"""
-Stacktrace cache
-TODO
-"""
-
-
-"""
-Collective Matching logic
-
-NOTE: For now, these collectives need to be supported by NCCL,
-https://docs.nvidia.com/deeplearning/nccl/user-guide/docs/overview.html.
-"""
-COLLECTIVES = {
-    "broadcast",
-    "all_gather",
-    "all_reduce",
-    "_all_gather_base",
-    "all_gather_into_tensor_coalesced",
-    "reduce_scatter_tensor_coalesced",
-    "_reduce_scatter_base",
-    "gather",
-    "scatter",
-    "alltoall_base",
-    "alltoall",
-}
-
-P2P = {
-    "send",
-    "recv",
-}
-
-
-class MatchState(Enum):
-    """
-    Enum representing the possible states of matching for collective operations.
-
-    - FULLY_MATCHED: Indicates that all aspects of the collective operations match.
-    - COLLECTIVE_TYPE_MISMATCH: The types of the collective operations differ.
-    - SIZE_OR_SYNTAX_MISMATCH: There is a mismatch in input/output sizes or violation of collective syntax.
-    - COLLECTIVE_STATE_MISMATCH:
-        The states of the collective not same, such as one finished while another just started or scheduled.
-    - UNDECIDED:
-        The match status is ambiguous or cannot be determined, e.g., we might need to check all ranks for alltoall_base.
-    """
-
-    FULLY_MATCHED = 1
-    COLLECTIVE_TYPE_MISMATCH = 2
-    SIZE_OR_SYNTAX_MISMATCH = 3
-    COLLECTIVE_STATE_MISMATCH = 4
-    UNDECIDED = 5
-
-
-def check_size_even_expand(list1: List[Any], list2: List[Any], size: int) -> bool:
-    ratio = None
-    for a, b in zip(list1, list2):
-        current_ratio = int(a) / int(b)
-        if current_ratio == 1:
-            continue
-        if current_ratio != size:
-            return False
-        elif ratio is None:
-            ratio = current_ratio
-        else:
-            return False
-    return True
-
-
-def check_size_alltoall(alltoall_cases: List[Dict[str, Any]]) -> Tuple[bool, int, int]:
-    input_numel = 0
-    output_numel = 0
-    for e in alltoall_cases:
-        input_numel += math.prod(e["input_sizes"][0])
-        output_numel += math.prod(e["output_sizes"][0])
-    return input_numel == output_numel, input_numel, output_numel
-
-
-class Op:
-    """Parses relevant info about operation out of 'event' dict
-
-    examples of supported `profiling_name`s:
-        nccl:broadcast
-        nccl:send 1->2
-        nccl:recv 3<-0
-    """
-
-    def __init__(self, event: Dict[Any, Any], memberships: Dict[str, Set[Any]]):
-        profiling_name = event["profiling_name"]
-        nccl, name = profiling_name.split(":")
-        assert nccl == "nccl", f"name formatting error? {nccl} != 'nccl'"
-        parts = name.split(" ")
-        type = parts[0]
-        meta = parts[1] if len(parts) == 2 else None
-        self.state = event["state"]
-
-        self.pg_name, _ = event["process_group"]
-
-        assert type in COLLECTIVES | P2P | {
-            "coalesced"
-        }, f"{type} is not a supported operation"
-        self.type = type
-        if type == "send":
-            s, d = meta.split("->")
-            self._src, self._dst = int(s), int(d)
-        elif type == "recv":
-            d, s = meta.split("<-")
-            self._dst, self._src = int(d), int(s)
-        else:
-            self._src, self._dst = -1, -1
-        pg_name, pg_desc = event["process_group"]
-        self._init_global_src_dst(memberships[pg_name])
-        self.pg_size = len(memberships[pg_name])
-
-        if type in P2P | COLLECTIVES:
-            self.input_sizes = event["input_sizes"]
-            self.output_sizes = event["output_sizes"]
-        else:
-            self.input_sizes, self.output_sizes = None, None
-        self.collective_seq_id = event["collective_seq_id"]
-        self.p2p_seq_id = event["p2p_seq_id"]
-
-    def _init_global_src_dst(self, pg_ranks: Set[Any]) -> None:
-        pg_ranks = sorted(pg_ranks)
-        self._src_g = pg_ranks[self._src] if self._src is not None else None
-        self._dst_g = pg_ranks[self._dst] if self._dst is not None else None
-
-    @property
-    def src(self) -> int:
-        assert self.type in P2P, "can't get src of non-p2p op"
-        return self._src
-
-    @property
-    def dst(self) -> int:
-        assert self.type in P2P, "can't get dst of non-p2p op"
-        return self._dst
-
-    def __repr__(self) -> str:
-        if self.type in P2P:
-            return (
-                f"{self.type}(s={self._src_g} d={self._dst_g}, sz={self.input_sizes})"
-            )
-        return f"{self.type}(input_sizes={self.input_sizes}, {self.state})"
-
-    def match(self, other: "Op") -> MatchState:
-        # TODO: I think this can validly not match,
-        # e.g. if one PG was used for p2p ops between only some of the peers?
-        # if self.seq_id != other.seq_id:
-        # return False
-
-        if self.type == "send":
-            # TODO: We need more states for p2p ops.
-            return (
-                MatchState.FULLY_MATCHED
-                if (
-                    other.type == "recv"
-                    and self.src == other.src
-                    and self.dst == other.dst
-                    and self.input_sizes == other.output_sizes
-                )
-                else MatchState.SIZE_OR_SYNTAX_MISMATCH
-            )
-        elif self.type == "recv":
-            return (
-                MatchState.FULLY_MATCHED
-                if (
-                    other.type == "send"
-                    and self.src == other.src
-                    and self.dst == other.dst
-                    and self.output_sizes == other.input_sizes
-                )
-                else MatchState.SIZE_OR_SYNTAX_MISMATCH
-            )
-        elif self.type in COLLECTIVES:
-            if self.type != other.type:
-                return MatchState.COLLECTIVE_TYPE_MISMATCH
-            if self.type in ["alltoall", "alltoall_base"]:
-                return MatchState.UNDECIDED
-            if self.type != "scatter" and self.input_sizes != other.input_sizes:
-                return MatchState.SIZE_OR_SYNTAX_MISMATCH
-            if self.type != "gather" and self.output_sizes != other.output_sizes:
-                return MatchState.SIZE_OR_SYNTAX_MISMATCH
-            if self.type == "all_reduce" and self.input_sizes != other.output_sizes:
-                return MatchState.SIZE_OR_SYNTAX_MISMATCH
-            # TODO: need to consider uneven sharding for all-gather.
-            # TODO: need to consider all_gather_into_tensor_coalesced (coalesced related)
-            if self.type in [
-                "all_gather",
-                "all_gather_base",
-            ] and not check_size_even_expand(
-                other.output_sizes, self.input_sizes, self.pg_size
-            ):
-                return MatchState.SIZE_OR_SYNTAX_MISMATCH
-            if self.type in [
-                "reduce_scatter",
-                "_reduce_scatter_base",
-            ] and not check_size_even_expand(
-                other.input_sizes, self.output_sizes, self.pg_size
-            ):
-                return MatchState.SIZE_OR_SYNTAX_MISMATCH
-            # TODO: need to add more checks for gather and scatter.
-            if self.state != other.state:
-                return MatchState.COLLECTIVE_STATE_MISMATCH
-        elif self.type == "coalesced":
-            return (
-                MatchState.FULLY_MATCHED
-                if (other.type == "coalesced")
-                else MatchState.SIZE_OR_SYNTAX_MISMATCH
-            )
-        return MatchState.FULLY_MATCHED
-
-
-def match_one_event(
-    event_a: Dict[Any, Any],
-    event_b: Dict[Any, Any],
-    memberships: Dict[str, Set[Any]],
-) -> MatchState:
-    op_a = Op(event_a, memberships)
-    op_b = Op(event_b, memberships)
-    return op_a.match(op_b)
-
-
-def match_coalesced_groups(
-    all_rank_events: Dict[Any, Any],
-    group_size: int,
-    groups: Dict[str, Group],
-    memberships: Dict[str, Set[Any]],
-) -> bool:
-    """
-    all_rank_events: {
-        rank: [
-            (idx, event_dict)
-        ]
-    }
-
-    Note: it is possible for event dicts in a coalesced group to be asymmetric.
-        e.g. the following events lists form a valid coalescing group
-             events0 [send:1]
-             events1 [recv:0, send:2]
-             events2 [recv:1]
-
-    Rule 1: all ops should find a match
-    Rule 2: relative ordering of sends and recvs in one event list can be arbitrary
-        e.g.
-        events1 [recv:0, send:2]  —> okay
-        events1 [send:2, recv:0] —> also okay
-    Rule 3: sends to the same dest or recvs from the src should be in a consistent order
-        e.g.
-        rank0 [send:1 (100B), send:1 (1000B)]
-        rank1 [recv:0 (1000B), recv:0 (100B)]   —> not okay
-    """
-    all_ops = {
-        rank: [Op(e, memberships) for i, e in all_rank_events[rank]]
-        for rank in all_rank_events
-    }
-
-    def visualize_ops(match: bool) -> None:
-        all_ops = {
-            rank: [Op(e, memberships) for i, e in all_rank_events[rank]]
-            for rank in all_rank_events
-        }
-
-        i = 0
-        row = []
-        progress = True
-        table = []
-        while progress:
-            progress = False
-            for r in all_ops:
-                if len(all_ops[r]) > i:
-                    _, event = all_rank_events[r][i]
-                    row.append(Op(event, memberships))
-                    progress = True
-                else:
-                    row.append(None)  # type: ignore[arg-type]
-            table.append(row)
-            row = []
-            i += 1
-        title = "Match" if match else "MISMATCH"
-        print(f"{title}\n", tabulate(table))  # type: ignore[operator]
-
-    # TODO can't verify seq_id bc there might have been valid seq deltas between ranks even within a pg.
-    for op_list in all_ops.values():
-        if not op_list:
-            # print("TODO- not sure if its valid for only some ranks in a PG to participate in a coalesced op?")
-            return False
-        assert op_list[-1].type == "coalesced"
-        op_list.pop(-1)
-
-    while all_ops:
-        first_rank = next(iter(all_ops))
-        my_ops = all_ops[first_rank]
-
-        if len(all_ops[first_rank]) == 0:
-            all_ops.pop(first_rank)
-            continue
-
-        # lets match the first collective! we need to know which ranks are involved, and ensure that this same
-        # collective is also the first one on those ranks within that group
-        op = my_ops[0]
-        match_idx = -1
-        if op.type in P2P:
-            dst_global_rank = sorted(memberships[op.pg_name])[op.dst]
-            peer_ops = all_ops[dst_global_rank]
-            for i, other in enumerate(peer_ops):
-                if op.match(other) == MatchState.FULLY_MATCHED:
-                    match_idx = i
-                    break
-                elif op.dst == other.src:
-                    # Rule 3
-                    break
-                else:
-                    # Rule 1
-                    continue
-        else:
-            raise NotImplementedError("coalesced collective ops")
-        if match_idx >= 0:
-            my_ops.pop(0)
-            peer_ops.pop(match_idx)
-        else:
-            visualize_ops(False)
-            return False
-
-    visualize_ops(True)
-    return True
-
-
 """
 Flat DB builder
 """
-
-
 def build_groups_memberships(
     pg_config: Any,
 ) -> Tuple[List[Group], Dict[Any, Group], List[Membership], Dict[str, Set[Any]]]:
@@ -530,57 +110,6 @@ def build_nccl_call(
         collective_type=entry["profiling_name"],
         sizes=entry["input_sizes"],
     )
-
-
-def find_coalesced_group(
-    pg_name: str, entries: List[Dict[str, Any]]
-) -> List[Tuple[int, Dict[str, Any]]]:
-    """Given a list of entries, if the collective_seq_id of the first entry matches that of subsequent ones,
-    build an return a list of entries terminating in a 'coalesced' op entry all sharing a collective_seq_id
-    TODO: handle p2p_seq_id v/s collective_seq_id separately here.
-    """
-    found = []
-    collective_seq_id = None
-    for i, e in enumerate(entries):
-        if e["process_group"][0] != pg_name:
-            continue
-        elif collective_seq_id is None:
-            collective_seq_id = e["collective_seq_id"]
-            found.append((i, e))
-        elif e["collective_seq_id"] == collective_seq_id:
-            found.append((i, e))
-        else:
-            break
-
-    if len(found) > 1:
-        assert found[-1][1]["profiling_name"] == "nccl:coalesced"
-        return found
-    return []
-
-
-def just_print_entries(
-    all_entries: Dict[int, List[Dict[str, Any]]],
-    _groups: Dict[str, Group],
-    _memberships: Dict[str, Set[Any]],
-) -> None:
-    rows = []
-    ranks = sorted(all_entries.keys())
-    headers = [f"Rank {rank}" for rank in ranks]
-    progress = True
-    while progress:
-        progress = False
-        row = []
-        for rank in ranks:
-            if len(all_entries[rank]) == 0:
-                row.append("")
-            else:
-                entry = all_entries[rank].pop(0)
-                row.append(str(Op(entry, _memberships)))
-                progress = True
-        if progress:
-            rows.append(row)
-
-    print(tabulate(rows, headers=headers))  # type: ignore[operator]
 
 
 def build_collectives(
@@ -821,39 +350,6 @@ def build_collectives(
             sys.exit(-1)
 
     return tracebacks, collectives, nccl_calls
-
-
-def check_no_missing_dump_files(
-    entries: Dict[str, Any], memberships: List[Membership]
-) -> None:
-    all_ranks = set()
-    for membership in memberships:
-        all_ranks.add(str(membership.global_rank))
-    dumps_ranks = set(entries.keys())
-    assert (
-        dumps_ranks == all_ranks
-    ), f"Missing dump files from ranks {all_ranks - dumps_ranks}"
-
-
-def check_version(versions: Dict[str, Any]) -> None:
-    for rank, version in versions.items():  # noqa: PERF102
-        major, minor = map(int, version.split("."))
-        # assert major == 2, f"Rank {rank} unsupported version {version}"
-        # assert minor >= 0, f"Rank {rank} unsupported version {version}"
-
-
-def check_trace_from_beginning(entries: Dict[str, Any]) -> bool:
-    for rank in entries:
-        first_record_id = entries[rank][0]["record_id"]
-        # TODO add more sequence information such that analysis can proceed even without complete buffer
-
-        # assert first_record_id == 0, f"Rank {rank} trace does not start at time 0 (first record is {first_record_id}."
-        if first_record_id != 0:
-            print(
-                f"Rank {rank} trace does not start at time 0 (first record is {first_record_id}."
-            )
-            return False
-    return True
 
 
 def build_db(details: Dict[str, Dict[str, Any]], args: argparse.Namespace) -> Database:
