+++ tools/flight_recorder/fr_trace.py
@@ -30,9 +30,10 @@ python fr_trace.py -d <dump dir containing trace files> [-o <output file>]
 import argparse
 import pickle
 
+from tools.flight_recorder.components.builder import build_db
 from tools.flight_recorder.components.config_manager import JobConfig
 from tools.flight_recorder.components.loader import read_dir
-from tools.flight_recorder.components.processor import build_db, types
+from tools.flight_recorder.components.types import types
 
 
 def main(args: argparse.Namespace) -> None:
